Gradle: Move SPDX utilities to their own module
To prepare for probably publishing these as a separate library.
