+++ model/src/main/kotlin/Package.kt
@@ -21,7 +21,7 @@ package com.here.ort.model
 
 import com.fasterxml.jackson.annotation.JsonInclude
 
-import com.here.ort.model.spdx.SpdxExpression
+import com.here.ort.spdx.SpdxExpression
 
 import java.util.SortedSet
