+++ settings.gradle
@@ -30,5 +30,6 @@ include 'model'
 include 'reporter'
 include 'reporter-web-app'
 include 'scanner'
+include 'spdx-utils'
 include 'test-utils'
 include 'utils'
