+++ spdx-utils/src/main/antlr/com/here/ort/spdx/SpdxExpression.g4
@@ -24,7 +24,7 @@
 grammar SpdxExpression;
 
 @header {
-package com.here.ort.model.spdx;
+package com.here.ort.spdx;
 }
 
 /*
