+++ spdx-utils/src/main/kotlin/SpdxLicenseException.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.model.spdx
+package com.here.ort.spdx
 
 /**
  * An enum containing all SPDX license exception IDs. This class is generated by the Gradle task
