+++ utils/src/test/kotlin/ExtensionsTest.kt
@@ -19,8 +19,6 @@
 
 package com.here.ort.utils
 
-import io.kotlintest.matchers.beEmpty
-import io.kotlintest.should
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldNotBe
 import io.kotlintest.shouldThrow
@@ -28,26 +26,11 @@ import io.kotlintest.specs.WordSpec
 
 import java.io.File
 import java.io.IOException
-import java.time.DayOfWeek
 
 class ExtensionsTest : WordSpec({
     "ByteArray.toHexString" should {
         "correctly convert a byte array to a string of hexadecimal digits" {
             byteArrayOf(0xde.toByte(), 0xad.toByte(), 0xbe.toByte(), 0xef.toByte()).toHexString() shouldBe "deadbeef"
-        }
-    }
-
-    "EnumSet.plus" should {
-        "create an empty set if both summands are empty" {
-            val sum = enumSetOf<DayOfWeek>() + enumSetOf()
-
-            sum should beEmpty()
-        }
-
-        "create the correct sum of two sets" {
-            val sum = enumSetOf(DayOfWeek.MONDAY, DayOfWeek.TUESDAY) + enumSetOf(DayOfWeek.TUESDAY, DayOfWeek.WEDNESDAY)
-
-            sum shouldBe enumSetOf(DayOfWeek.MONDAY, DayOfWeek.TUESDAY, DayOfWeek.WEDNESDAY)
         }
     }
